pyproject.toml
[tool.poetry]
name = "django-skeleton"
version = "0.1.0"
description = ""
authors = ["Eric Hanchrow <eric.hanchrow@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
pytest-django = "^4.8.0"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
# -- recommended but optional:
python_files = ["tests.py", "test_*.py", "*_test.py", "testing/python/*.py"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
